requirements.txt
google-genai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
Pillow>=10.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
crewai>=0.108.0
crewai-tools>=0.38.0
